synthetic_data_generator/requirements.txt
pymadcad==0.16.0
numpy-stl==3.0.1
pandas==2.1.1
